feat: implement meshing
This commit implements meshing by changing the way the pool allocator
gets pages. The trick is to use memfd_create to get an anonymous file
and mmap it (the SHARED flag seems to be required for meshing). Having
the file descriptor returned by memfd_create allows the virtual page
mappings to be updated by using mmap with MAP.FIXED to change the
physical address of a virtual page.

This commit does not do any page protecting/locking of pages during
meshing, so concurrent access is not supported.