refactor: dont move partial page pointer during alloc
Previous the current page for allocating slots was just the head of the
partial_pages list. This meant that if the slot allocated was the one
storing the list node, this node had to be moved to a new slot. By
storing a separate index for the current page for allocation simplifies
things by removing this complex bookkeeping.